move APIFnGet to action_get.go
